+++ packages/@ember/application/instance.js
@@ -6,7 +6,7 @@ import { assign } from 'ember-utils';
 import { get, set, computed } from 'ember-metal';
 import { environment } from 'ember-environment';
 import { jQuery } from 'ember-views';
-import { EngineInstance } from 'ember-application';
+import EngineInstance from '@ember/engine/instance';
 import { renderSettled } from 'ember-glimmer';
 
 /**
