--- packages/ember-application/index.js
@@ -1,1 +1,0 @@
-export { default as EngineInstance } from './lib/system/engine-instance';
