+++ packages/ember/index.js
@@ -100,7 +100,6 @@ import {
 import VERSION from './version';
 import * as views from 'ember-views';
 import * as routing from 'ember-routing';
-import * as application from 'ember-application';
 import * as extensionSupport from 'ember-extension-support';
 import EmberError from '@ember/error';
 import * as runloop from '@ember/runloop';
@@ -110,6 +109,7 @@ import Application, { onLoad, runLoadHooks } from '@ember/application';
 import Resolver from '@ember/application/globals-resolver';
 import ApplicationInstance from '@ember/application/instance';
 import Engine from '@ember/engine';
+import EngineInstance from '@ember/engine/instance';
 
 // ****ember-environment****
 
@@ -140,6 +140,7 @@ Ember.ApplicationInstance = ApplicationInstance;
 
 // ****@ember/engine****
 Ember.Engine = Engine;
+Ember.EngineInstance = EngineInstance;
 
 // ****ember-utils****
 Ember.generateGuid = utils.generateGuid;
@@ -508,9 +509,6 @@ Ember.generateController = routing.generateController;
 Ember.RouterDSL = routing.RouterDSL;
 Ember.Router = routing.Router;
 Ember.Route = routing.Route;
-
-// ****ember-application****
-Ember.EngineInstance = application.EngineInstance;
 
 runLoadHooks('Ember.Application', Application);
